Add beacon support
beacon cna file can be loaded on the team server or from cs. This cna
file will send a curl request to a keybase forwarder.
The keybase forwarder sets up an http listener and then performs
the network connection to the keybase api. The resulting message from
the connection will be sent to the channel specified in the cna file.

Signed-off-by: James Cook <jbcook@protonmail.com>